+++ work-order-backend/src/main/java/com/eeit219/work_order_system/modules/e/repository/NotificationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eeit219.work_order_system.modules.e.entity.Notification;
 
-public interface NotificationRepository extends JpaRepository<Notification, Long> {
+public interface NotificationRepository extends JpaRepository<Notification, Integer> {
     // 繼承 JpaRepository 後，自動擁有 save()、findById()、findAll() 等基本 CRUD 功能
-	List<Notification> findByReceiverIdOrderByNotificationIdDesc(Long receiverId);
+    List<Notification> findByReceiverIdOrderByNotificationIdDesc(Integer receiverId);
 }
